move Chi2ChargeMeasurement
