+++ RecoTracker/MeasurementDet/plugins/TkStripMeasurementDet.h
@@ -20,6 +20,7 @@
 #include "RecoTracker/TransientTrackingRecHit/interface/TSiStripRecHit2DLocalPos.h"
 
 #include "TrackingTools/DetLayers/interface/MeasurementEstimator.h"
+#include "RecoTracker/MeasurementDet/interface/ClusterFilterPayload.h"
 
 #include<tuple>
 
@@ -175,11 +176,11 @@ public:
     if (isMasked(*cluster)) return true;
     const GeomDetUnit& gdu( specificGeomDet());
     if (!accept(cluster, skipClusters)) return true;
+    if (!est.preFilter(ltp, ClusterFilterPayload(rawId(),&*cluster) )) return true;  // we know is on the cluster...
     VLocalValues const & vlv = cpe()->localParametersV( *cluster, gdu, ltp);
     bool isCompatible(false);
     for(auto vl : vlv) {
       SiStripRecHit2D recHit(vl.first, vl.second, fastGeomDet(), cluster);
-      if (!est.preFilter(ltp, recHit)) return true;  // we know is on the cluster... 
       std::pair<bool,double> diffEst = est.estimate(ltp, recHit);
       LogDebug("TkStripMeasurementDet")<<" chi2=" << diffEst.second;
       if ( diffEst.first ) {
@@ -198,11 +199,11 @@ public:
     if (isMasked(*cluster)) return true;
     const GeomDetUnit& gdu( specificGeomDet());
     if (!accept(cluster, skipClusters)) return true;
+    if (!est.preFilter(ltp, ClusterFilterPayload(rawId(),&*cluster) )) return true;
     VLocalValues const & vlv = cpe()->localParametersV( *cluster, gdu, ltp);
     bool isCompatible(false);
     for(auto vl : vlv) {
       auto && recHit  = SiStripRecHit2D( vl.first, vl.second, gdu, cluster);
-      if (!est.preFilter(ltp, recHit)) return true;  //	we know	is on the cluster...
       std::pair<bool,double> diffEst = est.estimate(ltp, recHit);
       LogDebug("TkStripMeasurementDet")<<" chi2=" << diffEst.second;
       if ( diffEst.first ) {
