+++ TrackingTools/KalmanUpdators/plugins/module.cc
@@ -7,12 +7,10 @@
 #include "FWCore/Framework/interface/ESHandle.h"
 #include "FWCore/Framework/interface/ModuleFactory.h"
 #include "FWCore/Framework/interface/ESProducer.h"
-#include "TrackingTools/KalmanUpdators/interface/Chi2ChargeMeasurementEstimatorESProducer.h"
 
 DEFINE_FWK_EVENTSETUP_MODULE(KFUpdatorESProducer);
 DEFINE_FWK_EVENTSETUP_MODULE(KFSwitching1DUpdatorESProducer);
 DEFINE_FWK_EVENTSETUP_MODULE(Chi2MeasurementEstimatorESProducer);
 DEFINE_FWK_EVENTSETUP_MODULE(MRHChi2MeasurementEstimatorESProducer);
 DEFINE_FWK_EVENTSETUP_MODULE(TrackingRecHitPropagatorESProducer);
-DEFINE_FWK_EVENTSETUP_MODULE(Chi2ChargeMeasurementEstimatorESProducer);
